add test file

+++ src/MLPnPsolver.cpp
@@ -125,7 +125,7 @@ namespace ORB_SLAM3 {
             vector<int> indexes(mRansacMinSet);
 
 	        // Get min set of points
-	        for(short i = 0; i < mRansacMinSet; ++i)
+	        for(int i = 0; i < mRansacMinSet; ++i)
 	        {
 	            int randi = DUtils::Random::RandomInt(0, vAvailableIndices.size()-1);
